US8659: move tests
