+++ SignInCheckIn/SignInCheckIn.Tests/SignInCheckIn.Tests.csproj
@@ -204,6 +204,7 @@
     <Compile Include="Controllers\MPAuthTestExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\ChildCheckinServiceTest.cs" />
+    <Compile Include="Services\FamilyServiceTest.cs" />
     <Compile Include="Services\LookupServiceTest.cs" />
     <Compile Include="Services\SiteServiceTest.cs" />
     <Compile Include="Services\WebsocketServiceTest.cs" />
